symlink

+++ packages/tools/src/utils.ts
@@ -2,7 +2,7 @@ import child from 'node:child_process';
 import { mkdir, writeFile as _writeFile, readFile as _readFile, unlink } from 'fs/promises';
 import { dirname, join, relative, resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
-import { existsSync, statSync } from 'node:fs';
+import { statSync } from 'node:fs';
 
 export const dirnameForFileURL = (url: string) => dirname(fileURLToPath(url));
 
@@ -20,21 +20,35 @@ export const exec = (command: string, cwd?: string) => {
       if (code !== 0) {
         return reject(`exec failed with code ${code}`);
       }
-      console.log();
       resolve(0);
     });
   });
 };
 
+export const statOptional = (path: string) => {
+  try {
+    return statSync(path);
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  } catch(err: any) {
+    if(err.code === 'ENOENT') {
+      return undefined;
+    }
+    throw err;
+  }
+}
+
+export const exists = (opts: { target: string, path: string }) => {
+  const stat = statOptional(join(opts.target, opts.path));
+  return !!stat;
+}
+
 export const symlink = async (opts: { source: string; target: string; }) => {
   const target = resolve(opts.target);
   const source = relative(dirname(target), resolve(opts.source));
-
-  const stat = statSync(target);
-  if(!stat.isSymbolicLink()) {
+  const stat = statOptional(target);
+  if(stat && !stat.isSymbolicLink()) {
     await unlink(target);
   }
-
   await mkdirp(dirname(target));
   await exec(`ln -s "${source}" "${target}"`);
 }
@@ -54,6 +68,7 @@ export const mkdirp = async (path: string) => {
 };
 
 export const writeString = async (path: string, data: string) => {
+  await mkdirp(dirname(path));
   await _writeFile(path, data, 'utf-8');
 };
